Fixed bugs and some code refactoring

+++ src/main/java/com/task/tradingPlatform/service/userService/UserServiceImpl.java
@@ -1,9 +1,13 @@
 package com.task.tradingPlatform.service.userService;
 
+import com.task.tradingPlatform.exception.DeleteYourAccountException;
 import com.task.tradingPlatform.exception.UserNotExistException;
 import com.task.tradingPlatform.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -12,11 +16,15 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserByEmail(String email) {
-        userRepository.deleteByEmail(email);
+        if (isUserExist(email)) {
+            userRepository.deleteByEmail(email);
+            if (Objects.equals(SecurityContextHolder.getContext().getAuthentication().getName(), email)) {
+                throw new DeleteYourAccountException("Вы удалили свой аккаунт и вышли из системы");
+            }
+        }
     }
 
-    @Override
-    public Boolean isUserExist(String email) {
+    private Boolean isUserExist(String email) {
         userRepository.findByEmail(email).orElseThrow(
                 () -> new UserNotExistException("Пользователя c email: " + email + " не существует")
         );
